requirements.txt
pandas
openpyxl
httpx
python-dotenv
requests
sentence-transformers
scikit-learn
streamlit
